refactor: move itens dentro de common/frontend para frontend/common
