+++ src/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable/modules/RenderDateTagCell.tsx
@@ -1,4 +1,4 @@
-import { TableTag } from "@/features/acf/common/frontend/TableTag";
+import { TableTag } from "@/features/acf/frontend/common/TableTag";
 import React from "react";
 import { iconDetailsMap } from "../consts";
